requirements.txt
Flask==2.3.3
twilio==8.10.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
supabase>=1.0.3
psycopg2-binary>=2.9.5
python-crontab>=2.7.1
Pillow>=10.0.0